requirements.txt
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
scikit-learn==0.23.2
scipy==1.5.2
streamlit==0.65.0

